Fix issue with nvr_cameraPoll being private
A bad version of the code was uploaded with nvr_cameraPoll set to private, which would prevent it from being called within the runIn context.  Essentially, polling stopped.  If this was running on your hub just update the code, open the NVR smart app and hit "Done" which will cause the camera device to bootstrap itself back to normal.