Add NVR poll API failure detection
When the NVR API call for polling fails successively, set the camera connection status to disconnected. Dynamically change the poll interval up to a minute based on connection status to be a good citizen if the system is not responding. Add a check in the take() call to only send the API if we have a connection status of connected.