Send username and password to bootstrap instead of API key.
Workaround for reports that 3.1.x doesn't allow the API key to bootstrap.  Since the /api/2.0/login call will return the API key, store that and the device handler remains untouched for now.  Using the API key thereafter has the nice benefit that subsequent logins to the web portal with the same username and password won't start log out ST and vice-versa.