Add watchdog for polling & some hardening
SmartThings has a very unreliable "runIn()" implementation where sometimes it just doesn't get scheduled (or it just doesn't run).  Add a watchdog by using the "runEvery...()" context, which is a different type of scheduler and so far has been reliable.  

Harden the nvr_cameraPollCallback function against missing data during errors (rare, usually happens only when an update() happens in the middle of a request).

Fixup update() to not artificially bang the connection and motion status.