Update README.md
UniFi Protect is in unmaintained status and not recommended for use.  Ubiquiti changed the login architecture from local to cloud, which makes this hard to make work, or just not possible with the current implementation. The project is now closed.

UniFi NVR is in unmaintained status as the NVR software has been abandoned by Ubiquiti. It might still work for the foreseeable future, but I will not be committing any further effort. The project is now closed.